+++ docs/realm.js
@@ -396,7 +396,7 @@ class Realm {
  *   user id in order to avoid collisions with other apps or users. An absolute path is left untouched
  *   and on some platforms (iOS and Android) the app might not have permissions to create or open
  *   the file - permissions are not validated. If a relative path is specified, it is relative to
- *   {@link Realm.App~AppConfiguration.syncRootDirectory}.
+ *   {@link Realm.App~AppConfiguration.baseFilePath}.
  * @property {string} [fifoFilesFallbackPath] - Opening a Realm creates a number of FIFO special files in order to
  * coordinate access to the Realm across threads and processes. If the Realm file is stored in a location
  * that does not allow the creation of FIFO special files (e.g. FAT32 filesystems), then the Realm cannot be opened.