+++ src/js_app.hpp
@@ -171,7 +171,7 @@ void AppClass<T>::constructor(ContextType ctx, ObjectType this_object, Arguments
     static const String config_app = "app";
     static const String config_app_name = "name";
     static const String config_app_version = "version";
-    static const String config_sync_root_directory = "syncRootDirectory";
+    static const String config_base_file_path = "baseFilePath";
 
     args.validate_count(1);
 
@@ -224,9 +224,9 @@ void AppClass<T>::constructor(ContextType ctx, ObjectType this_object, Arguments
             }
         }
 
-        ValueType sync_root_directory_value = Object::get_property(ctx, config_object, config_sync_root_directory);
-        if (!Value::is_undefined(ctx, sync_root_directory_value)) {
-            client_config.base_file_path = Value::validated_to_string(ctx, sync_root_directory_value);
+        ValueType base_file_path_value = Object::get_property(ctx, config_object, config_base_file_path);
+        if (!Value::is_undefined(ctx, base_file_path_value)) {
+            client_config.base_file_path = Value::validated_to_string(ctx, base_file_path_value);
         }
         else {
             client_config.base_file_path = default_realm_file_directory();